requirements.txt
aiohttp==3.8.6
pydantic==2.5.0
pydantic-settings==2.1.0
sentry-sdk==1.35.0
PyYAML==6.0.1
plumbum==1.8.2
pytest==7.4.3
pytest-cov==4.1.0
anyio==4.0.0
